v0.1.14, a somewhat working collision system

+++ include/player.h
@@ -25,8 +25,8 @@ public:
     //player assets and simialr stuff
     Vector2 position = { 0.0f, 0.0f };
     Texture2D model_movement = LoadTexture("../img/player/Unarmed_Movement_full2.png");
-    Rectangle frameRecMove = { 0.0f, 0.0f, model_movement.width/8, model_movement.height/3 };
-    Rectangle frameRecIdle = { 0.0f, 0.0f, model_movement.width/8, model_movement.height/3 };
+    Rectangle frameRecMove = { 0.0f, 0.0f, (float)model_movement.width/8, (float)model_movement.height/3 };
+    Rectangle frameRecIdle = { 0.0f, 0.0f, (float)model_movement.width/8, (float)model_movement.height/3 };
     PlayerState state = GROUND;
     Rectangle hitbox = { position.x, position.y, frameRecIdle.width, frameRecIdle.height };
